feat(Structures): Added SiteDto and started implementing gridView for site costs

+++ src/main/java/de/uniks/stp24/component/menu/SitePropertiesComponent.java
@@ -5,6 +5,8 @@ import de.uniks.stp24.model.Resource;
 import de.uniks.stp24.service.ResourcesService;
 import de.uniks.stp24.service.TokenStorage;
 import de.uniks.stp24.service.game.IslandsService;
+import javafx.collections.FXCollections;
+import javafx.collections.ObservableList;
 import javafx.fxml.FXML;
 import javafx.scene.control.Button;
 import javafx.scene.control.ListView;
@@ -15,6 +17,7 @@ import javafx.scene.text.TextAlignment;
 import org.controlsfx.control.GridCell;
 import org.controlsfx.control.GridView;
 import org.fulib.fx.annotation.controller.Component;
+import org.fulib.fx.annotation.event.OnInit;
 import org.fulib.fx.annotation.event.OnRender;
 import org.fulib.fx.controller.Subscriber;
 
@@ -29,7 +32,7 @@ public class SitePropertiesComponent extends AnchorPane {
     @FXML
     ListView<Resource> siteConsumesListView;
     @FXML
-    GridView<Resource> siteCostsGridView;
+    GridView<Map<String, Integer>> siteCostsGridView;
     @FXML
     Button buildSiteButton;
     @FXML
@@ -64,6 +67,8 @@ public class SitePropertiesComponent extends AnchorPane {
     }
     private Island island;
 
+    public ObservableList<Map<String, Integer>> resources = FXCollections.observableArrayList();
+
     @OnRender
     public void render(){
         this.siteType = "energy";
@@ -86,26 +91,31 @@ public class SitePropertiesComponent extends AnchorPane {
     }
 
     public void destroySite(){
-        island = tokenStorage.getIsland();
-        subscriber.subscribe(resourcesService.destroySite(tokenStorage.getGameId(), island, siteType), result -> {
-            tokenStorage.setIsland(islandsService.updateIsland(result));
-        });
+        if (tokenStorage.getIsland().sites().get(siteType) != 0){
+            island = tokenStorage.getIsland();
+            subscriber.subscribe(resourcesService.destroySite(tokenStorage.getGameId(), island, siteType), result -> {
+                tokenStorage.setIsland(islandsService.updateIsland(result));
+            });
+        }
     }
-
+    @OnRender
     public void displayCostsOfSite(){
-        
+        subscriber.subscribe(resourcesService.getResourcesSite(siteType), result -> {
+            this.siteCostsGridView.setCellFactory(gridViewCells -> new CustomGridCell());
+            System.out.println(result.cost());
+        });
     }
 }
 
-class CustomGridCell extends GridCell<Resource> {
+class CustomGridCell extends GridCell<Map<String, Integer>> {
     @Override
-    protected void updateItem(Resource item, boolean empty) {
+    protected void updateItem(Map<String, Integer> item, boolean empty) {
         super.updateItem(item, empty);
         if (empty || item == null) {
             setText(null);
             setGraphic(null);
         } else {
-            setText(String.valueOf(item.count()));
+            setText(item.toString());
             setTextAlignment(TextAlignment.CENTER);
 
         }
